feat: 코인 슬롯 머신 페이지와 api 연동
