+++ main/main.html
@@ -9,6 +9,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Single+Day&display=swap" rel="stylesheet">
+    <script defer src="../upbitAPI/coin.js"></script>
     <script defer src="main.js"></script>
     <title>UselessCoinPick</title>
 </head>
